refactor: [#31] move module
